refactor(role): move `role` to separate file
