+++ src/action.cc
@@ -1,5 +1,7 @@
 #include "action.hh"
 
+#include <string>
+
 Action::Action(std::string name, bool has_target) {
     this->name = name;
     this->has_target = has_target;
